fix: prod build not compiling
