+++ app/src/main/kotlin/com/wire/android/WireApplication.kt
@@ -114,7 +114,7 @@ class WireApplication : Application(), Configuration.Provider {
 
     private fun initializeApplicationLoggingFrameworks() {
         // 1. Datadog should be initialized first
-        initDatadogLogger(this)
+        ExternalLoggerManager.initDatadogLogger(this)
         // 2. Initialize our internal logging framework
         appLogger = KaliumLogger(
             config = KaliumLogger.Config(
