--- app/src/prod/kotlin.com.wire.android/ExternalLoggerManager.kt
@@ -1,2 +1,0 @@
-package com.wire.android
-fun initDatadogLogger(context: Context) = Unit
