+++ app/src/prod/kotlin/com/wire/android/ExternalLoggerManager.kt
@@ -1,0 +1,7 @@
+package com.wire.android
+
+import android.content.Context
+
+object ExternalLoggerManager {
+    fun initDatadogLogger(context: Context) = Unit
+}
